Added separate comp for community sidebar
